testing improvements (#91)
* initial commit

* added react17 and 18 as peer dependencies

* only react18 as peer dependency

* update prettier and eslint scripts

* react as a peer dependency and test script improvements

* temporarily remove depcheck

* fix vite entry point for build

* started unit tests for core and root

* build options changes

* update github workflow

* regenerate package-lock.json

* misc test update

* improved build and testing

* clean ups and legacy export

* make types accessible in build

* restructure testing folders

* path fix;

* another path fix

* creared node scripts for copying test file

* update path for test copy script

* remove unneccesary root tests and update core tests

* remove can-stache test and npm package and wrap tests in describe block

* vite build test

* update typecheck script

* update tsconfig to exclude builds test folder

* update buildTests to build vite example build

* swc build

* webpack build test

* misc script fix

* buildtests fix

* clean up
