+++ package.json
@@ -3,6 +3,7 @@
   "version": "2.0.0-alpha.3",
   "description": "Convert react components to native Web Components.",
   "type": "module",
+  "types": "dist/root/root.d.ts",
   "exports": {
     ".": {
       "import": "./dist/root.js",
@@ -36,8 +37,6 @@
     "@types/react": "^18.0.33",
     "@types/react-dom": "^18.0.11",
     "@webcomponents/custom-elements": "^1.2.4",
-    "can-stache": "^4.17.20",
-    "can-stache-bindings": "^4.10.9",
     "depcheck": "^1.4.3",
     "eslint": "^8.0.0",
     "eslint-plugin-react": "^7.21.4",
@@ -53,9 +52,9 @@
     "build": "vite build",
     "ci": "npm run test",
     "buildtests": "chmod +x src/tests/buildtests && ./src/tests/buildtests",
-    "test": "vitest src/tests/**/react-to-webcomponent.test.jsx && npm run test:core && npm run test:root",
+    "test": "vitest src/tests/environments/**/react-to-webcomponent.test.jsx && npm run test:core && npm run test:builds",
     "test:core": "vitest src/core/core.test.tsx",
-    "test:root": "vitest src/root/root.test.tsx",
+    "test:builds": "vitest src/tests/builds/builds.test.ts",
     "start": "http-server -c-1",
     "typecheck": "tsc --noEmit",
     "eslint": "eslint ./src/**/*.ts",
