+++ src/tests/builds/vite/vite.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    lib: {
+      entry: 'src/main.tsx',
+      name: 'App-WC',
+      fileName: 'index',
+      formats: ['es', 'cjs']
+    }
+  }
+})
