+++ src/tests/builds/webpack/.babelrc
@@ -1,0 +1,12 @@
+{
+  "presets": [
+    "@babel/preset-env",
+    "@babel/typescript",
+    [
+      "@babel/preset-react",
+      {
+        "runtime": "automatic"
+      }
+    ]
+  ]
+}
