+++ src/tests/builds/webpack/src/App.tsx
@@ -1,0 +1,7 @@
+import { FC } from "react"
+
+const App: FC<{ text: string }> = ({ text = 'hello, world' }) => {
+  return <div>{text}</div>
+}
+
+export default App
