+++ src/tests/builds/webpack/src/index.tsx
@@ -1,0 +1,7 @@
+import r2wc from "react-to-webcomponent"
+import App from "./App"
+
+const AppWC = r2wc(App, { props: ["text"] })
+
+export default AppWC
+
