+++ src/tests/buildtests
@@ -1,32 +1,39 @@
 #!/bin/bash
 
-
-sed -e "s/.\/legacy\/react-to-webcomponent/react-to-webcomponent\/render/g" -e "s/.*react-dom.*/import ReactDOM from 'react-dom'/g" src/react-to-webcomponent.test.jsx > src/tests/react16/react-to-webcomponent.test.jsx
-
-cd src/tests/react16
+cd src/tests/environments/react16
+node copyTest.js
 npm i
-npm install --install-links ../../..
-cd ../../../
+npm install --install-links ../../../..
+cd ../../../../
 
-sed -e "s/.\/legacy\/react-to-webcomponent/react-to-webcomponent\/render/g" -e "s/.*react-dom.*/import ReactDOM from 'react-dom'/g" src/react-to-webcomponent.test.jsx > src/tests/react17/react-to-webcomponent.test.jsx
-
-cd src/tests/react17
+cd src/tests/environments/react17
+node copyTest.js
 npm i
-npm install --install-links ../../..
-cd ../../../
+npm install --install-links ../../../..
+cd ../../../../
 
-
-sed "s/.\/legacy\/react-to-webcomponent/react-to-webcomponent/g" src/react-to-webcomponent.test.jsx > src/tests/react18/react-to-webcomponent.test.jsx
-
-cd src/tests/react18
+cd src/tests/environments/react18
+node copyTest.js
 npm i
-npm install --install-links ../../..
-cd ../../../
+npm install --install-links ../../../..
+cd ../../../../
 
-
-# sed -e 's/.\/react-to-webcomponent/react-to-webcomponent/g' -e 's/.*react-dom.*/import React from "preact\/compat"/g' -e '/import React from "react"/d' -e 's/ReactDOM/React/g' src/react-to-webcomponent.test.jsx > src/tests/preact10/react-to-webcomponent.test.jsx
-
-# cd src/tests/preact10
+# cd src/tests/environments/preact10
 # npm i
-# npm install --install-links ../../..
-# cd ../../../
+# npm install --install-links ../../../..
+# cd ../../../../
+
+
+# builds
+cd src/tests/builds/vite
+npm i
+npm install --install-links ../../../..
+npm run build
+node swcBuild.js
+cd ../../../../
+
+cd src/tests/builds/webpack
+npm i
+npm install --install-links ../../../..
+npm run build
+cd ../../../../
