+++ src/tests/environments/preact10/package.json
@@ -2,8 +2,6 @@
   "name": "react-to-webcomponent-preact10",
   "version": "0.0.0",
   "dependencies": {
-    "can-stache": "^4.17.20",
-    "can-stache-bindings": "^4.10.9",
     "prop-types": "^15.7.2",
     "preact": "^10.10.0",
     "react-to-webcomponent": "file:.."
