+++ src/tests/environments/react17/package.json
@@ -2,16 +2,14 @@
   "name": "react-to-webcomponent-react17",
   "version": "0.0.0",
   "dependencies": {
-    "can-stache": "^4.17.20",
-    "can-stache-bindings": "^4.10.9",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-to-webcomponent": "file:.."
+    "react-to-webcomponent": "file:../../../.."
   },
   "devDependencies": {
-    "typescript": "^4.3.5",
     "jsdom": "^20.0.0",
+    "typescript": "^4.3.5",
     "vite": "^3.0.3"
   }
 }
