+++ src/tests/environments/react18/package.json
@@ -2,12 +2,10 @@
   "name": "react-to-webcomponent-react18",
   "version": "0.0.0",
   "dependencies": {
-    "can-stache": "^4.17.20",
-    "can-stache-bindings": "^4.10.9",
     "prop-types": "^15.7.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-to-webcomponent": "file:../../.."
+    "react-to-webcomponent": "file:../../../.."
   },
   "devDependencies": {
     "jsdom": "^20.0.0",
