+++ tsconfig.json
@@ -21,5 +21,5 @@
     "skipLibCheck": true
   },
   "include": ["@types", "src"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "src/tests/builds"]
 }
